requirements.txt
streamlit
langchain
langgraph
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
langchain_community